spec fixes
Update metered_solar.rb
